backend/main.py
# backend/main.py - Run with uvicorn main:app
from fastapi import FastAPI
from pydantic import BaseModel
import json
import os
from pinecone import Pinecone
import psycopg2
from psycopg2.extras import RealDictCursor

app = FastAPI()

# Environment variables for production
PINECONE_API_KEY = os.getenv("PINECONE_API_KEY", "pcsk_5NXuE6_bfXMwA1Lj5djvEPPKqgN17FbicWVGEekcAxKiiouNPvjJrgnvDUSrtzXY9foE8")
PINECONE_INDEX_NAME = os.getenv("PINECONE_INDEX_NAME", "ivylevel-db")

# Real RDS Database setup from environment variables
DB_CONFIG = {
    'dbname': os.getenv("DB_NAME", "ivylevel"),
    'user': os.getenv("DB_USER", "ivylevel_admin"),
    'password': os.getenv("DB_PASSWORD", "YOUR_DATABASE_PASSWORD"),
    'host': os.getenv("DB_HOST", "ivylevel-database-1.cbeyu4q002ay.us-east-1.rds.amazonaws.com"),
    'port': int(os.getenv("DB_PORT", "5432"))
}

# Initialize Pinecone with new API
pc = Pinecone(api_key=PINECONE_API_KEY)
index = pc.Index(PINECONE_INDEX_NAME)

# Mock database for development
mock_users = []

# Database connection function
def get_db_connection():
    try:
        conn = psycopg2.connect(**DB_CONFIG)
        return conn
    except Exception as e:
        print(f"Database connection error: {e}")
        print("Using mock database for development")
        return None

# Mock implementations for services not set up yet
class MockSentenceTransformer:
    def encode(self, text):
        # Mock embedding - replace with real sentence-transformers later
        return [0.1] * 384  # 384 dimensions for all-MiniLM-L6-v2

class MockGraph:
    def invoke(self, data):
        # Mock multi-agent workflow - replace with real LangGraph later
        profile = data.get('profile', {})
        gpa = profile.get('gpa', 0)
        
        # Mock assessment logic
        if gpa >= 3.8:
            return {"chettyScore": 8.5, "path": "custom_10_days"}
        elif gpa >= 3.5:
            return {"chettyScore": 7.2, "path": "standard_14_days"}
        else:
            return {"chettyScore": 6.0, "path": "extended_21_days"}

# Initialize mock services
model = MockSentenceTransformer()
graph = MockGraph()

# Request models
class ProfileRequest(BaseModel):
    profile: dict

class QueryRequest(BaseModel):
    query: str

@app.post("/provision")  # For CUJ 6
def provision_coach(request: ProfileRequest):
    result = graph.invoke({"profile": request.profile})
    
    # Try to store in real database, fallback to mock
    conn = get_db_connection()
    if conn:
        try:
            cur = conn.cursor()
            cur.execute(
                "INSERT INTO users (id, role, profile) VALUES (%s, %s, %s) ON CONFLICT DO NOTHING",
                (f"user_{len(result)}", "coach", json.dumps(result))
            )
            conn.commit()
            cur.close()
            conn.close()
            print("✅ Stored in real database")
        except Exception as e:
            print(f"Database error: {e}")
            # Fallback to mock storage
            mock_users.append({
                "id": f"user_{len(mock_users)}",
                "role": "coach",
                "profile": result
            })
            print("✅ Stored in mock database")
    else:
        # Use mock storage
        mock_users.append({
            "id": f"user_{len(mock_users)}",
            "role": "coach",
            "profile": result
        })
        print("✅ Stored in mock database")
    
    return result

@app.post("/kb/query")  # For CUJ 9
def query_kb(request: QueryRequest):
    try:
        # Use real Pinecone for KB queries
        embedding = model.encode(request.query)
        results = index.query(vector=embedding, top_k=5)
        return results
    except Exception as e:
        # Fallback to mock results if Pinecone fails
        return {
            "matches": [
                {"id": "doc1", "score": 0.9, "metadata": {"content": "Sample KB content"}},
                {"id": "doc2", "score": 0.8, "metadata": {"content": "Another KB document"}}
            ]
        }

@app.get("/health")
def health():
    # Test database connection
    db_status = "disconnected"
    try:
        conn = get_db_connection()
        if conn:
            cur = conn.cursor()
            cur.execute("SELECT 1")
            cur.fetchone()
            cur.close()
            conn.close()
            db_status = "connected"
        else:
            db_status = "using_mock"
    except Exception as e:
        db_status = f"error: {str(e)}"
    
    return {
        "status": "healthy", 
        "pinecone": "connected", 
        "database": db_status,
        "mock_users_count": len(mock_users),
        "environment": os.getenv("ENVIRONMENT", "development"),
        "endpoints": {
            "provision": "/provision",
            "kb_query": "/kb/query",
            "health": "/health",
            "db_test": "/db/test"
        }
    }

@app.get("/")
def root():
    return {
        "message": "IvyLevel Backend API", 
        "version": "2.2",
        "environment": os.getenv("ENVIRONMENT", "development")
    }

@app.get("/db/test")
def test_database():
    """Test database connection and basic operations"""
    conn = get_db_connection()
    if not conn:
        return {
            "status": "using_mock",
            "database": "mock",
            "user_count": len(mock_users),
            "message": "Real database not accessible, using mock database"
        }
    
    try:
        cur = conn.cursor(cursor_factory=RealDictCursor)
        
        # Test basic query
        cur.execute("SELECT COUNT(*) as user_count FROM users")
        result = cur.fetchone()
        
        cur.close()
        conn.close()
        
        return {
            "status": "success",
            "database": "connected",
            "user_count": result['user_count'] if result else 0
        }
    except Exception as e:
        return {"status": "error", "message": str(e)}

@app.get("/mock/users")
def get_mock_users():
    """Get users from mock database (for development)"""
    return {
        "status": "success",
        "database": "mock",
        "users": mock_users,
        "count": len(mock_users)
    } 