backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pinecone==7.3.0
psycopg2-binary==2.9.10
pydantic==2.5.0
python-multipart==0.0.6 